+++ storage_utils/tests/test_swift.py
@@ -1881,6 +1881,22 @@ class TestStat(SwiftTestCase):
             'Access-Control': {
                 'read-only': ['seq_upload_rnd', 'swft_labprod'],
                 'read-write': ['svc_svc_seq']
+            },
+            'headers': {
+                'content-length': '0',
+                'x-account-storage-policy-3xreplica-container-count': '31',
+                'x-account-object-count': '20081986',
+                'connection': 'close',
+                'x-timestamp': '1445629170.46005',
+                'x-account-access-control': '{"read-only":["seq_upload_rnd","swft_labprod"],"read-write":["svc_svc_seq"]}',  # nopep8
+                'x-account-storage-policy-3xreplica-bytes-used': '24993077101523',  # nopep8
+                'x-trans-id': 'tx2acc1bc870884a0487dd0-0056a6a993',
+                'date': 'Mon, 25 Jan 2016 23:02:43 GMT',
+                'x-account-bytes-used': '24993077101523',
+                'x-account-container-count': '31',
+                'content-type': 'text/plain; charset=utf-8',
+                'accept-ranges': 'bytes',
+                'x-account-storage-policy-3xreplica-object-count': '20081986'
             }
         })
 
@@ -1925,7 +1941,22 @@ class TestStat(SwiftTestCase):
             'Containers-in-policy-"3xreplica"': '31',
             'Objects-in-policy-"3xreplica"': '20081986',
             'Bytes-in-policy-"3xreplica"': '24993077101523',
-            'Access-Control': {}
+            'Access-Control': {},
+            'headers': {
+                'content-length': '0',
+                'x-account-storage-policy-3xreplica-container-count': '31',
+                'x-account-object-count': '20081986',
+                'connection': 'close',
+                'x-timestamp': '1445629170.46005',
+                'x-account-storage-policy-3xreplica-bytes-used': '24993077101523',  # nopep8
+                'x-trans-id': 'tx2acc1bc870884a0487dd0-0056a6a993',
+                'date': 'Mon, 25 Jan 2016 23:02:43 GMT',
+                'x-account-bytes-used': '24993077101523',
+                'x-account-container-count': '31',
+                'content-type': 'text/plain; charset=utf-8',
+                'accept-ranges': 'bytes',
+                'x-account-storage-policy-3xreplica-object-count': '20081986'
+            }
         })
 
     def test_container(self):
@@ -1966,7 +1997,19 @@ class TestStat(SwiftTestCase):
             'Read-ACL': '',
             'Write-ACL': '',
             'Sync-To': '',
-            'Sync-Key': ''
+            'Sync-Key': '',
+            'headers': {
+                'content-length': '0',
+                'x-container-object-count': '43868',
+                'accept-ranges': 'bytes',
+                'x-storage-policy': '3xReplica',
+                'date': 'Mon, 25 Jan 2016 23:10:45 GMT',
+                'connection': 'close',
+                'x-timestamp': '1452627422.60776',
+                'x-trans-id': 'tx441a691b0e514782b51be-0056a6ab75',
+                'x-container-bytes-used': '55841489571',
+                'content-type': 'text/plain; charset=utf-8'
+            }
         })
 
     def test_object(self):
@@ -2012,7 +2055,23 @@ class TestStat(SwiftTestCase):
             'Content-Length': '112',
             'Last-Modified': 'Fri, 15 Jan 2016 05:22:46 GMT',
             'ETag': '87f0b7f04557315e6d1e6db21742d31c',
-            'Manifest': None
+            'Manifest': None,
+            'headers': {
+                'content-length': '112',
+                'x-object-meta-x-agi-ctime': '2016-01-15T05:22:00.0Z',
+                'x-object-meta-x-agi-mode': '436',
+                'accept-ranges': 'bytes',
+                'last-modified': 'Fri, 15 Jan 2016 05:22:46 GMT',
+                'connection': 'close',
+                'x-object-meta-x-agi-gid': '0',
+                'x-timestamp': '1452835365.34322',
+                'etag': '87f0b7f04557315e6d1e6db21742d31c',
+                'x-trans-id': 'tx805b2e7ce56343a6b2ea3-0056a6ac39',
+                'date': 'Mon, 25 Jan 2016 23:14:01 GMT',
+                'content-type': 'application/octet-stream',
+                'x-object-meta-x-agi-uid': '0',
+                'x-object-meta-x-agi-mtime': 'Fri, 15 Jan 2016 05:22:01 PST'
+            }
         })
 
     def test_getsize_content_length(self):