+++ docs/release_notes.rst
@@ -1,6 +1,12 @@
 Release Notes
 =============
 
+v6.9.0
+------
+
+* Added ``object_headers`` option to `SwiftPath.upload`
+* Headers are now returned in `SwiftPath.stat`
+
 v6.8.0
 ------